resturant_API/requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
coverage==6.3.2
Deprecated==1.2.13
Django==3.2
isort==4.3.13
flake8==3.7.7
pycodestyle==2.5.0
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
model-bakery==1.5.0
packaging==21.3
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
sqlparse==0.4.2
wrapt==1.14.0
